app/requirements.txt
Django==3.2.16
djangorestframework==3.13.1
drf-spectacular==0.26.0
uWSGI==2.0.21
psycopg2-binary==2.9.5
django-webpack-loader==1.7.0
requests
debugpy
django-health-check
django_redis
django-permissions-policy==4.14.0
django-cors-headers==3.13.0
django-csp==3.6
django-materializecss-form
python-magic
